Implement GraphQL Codegen with client preset
