+++ apps/store/.eslintrc.js
@@ -4,7 +4,7 @@ const baseConfig = require('eslint-config-custom')
 module.exports = {
   root: true,
   ...baseConfig,
-  ignorePatterns: ['src/services/graphql/*', '*.js'],
+  ignorePatterns: ['src/services/apollo/generated.ts', 'src/services/apollo/gql/*', '*.js'],
   parserOptions: {
     project: ['./tsconfig.json'],
     tsconfigRootDir: __dirname,
