+++ apps/store/package.json
@@ -28,6 +28,7 @@
     "@emotion/is-prop-valid": "1.2.1",
     "@emotion/react": "11.11.1",
     "@emotion/styled": "11.11.0",
+    "@graphql-typed-document-node/core": "3.2.0",
     "@lottiefiles/react-lottie-player": "3.5.3",
     "@radix-ui/react-accordion": "1.1.2",
     "@radix-ui/react-collapsible": "1.0.3",
@@ -82,6 +83,8 @@
     "@datadog/datadog-ci": "2.25.0",
     "@emotion/babel-plugin": "11.11.0",
     "@graphql-codegen/cli": "5.0.0",
+    "@graphql-codegen/client-preset": "4.1.0",
+    "@graphql-codegen/client-preset-swc-plugin": "0.2.0",
     "@graphql-codegen/typescript": "4.0.1",
     "@graphql-codegen/typescript-operations": "4.0.1",
     "@graphql-codegen/typescript-react-apollo": "4.1.0",
