--- apps/store/src/features/widget/ShopSessionCreatePartner.graphql
@@ -1,5 +1,0 @@
-mutation ShopSessionCreatePartner($input: ShopSessionCreatePartnerInput!) {
-  shopSessionCreatePartner(input: $input) {
-    id
-  }
-}
