+++ apps/store/src/features/widget/StartFlow.helpers.ts
@@ -1,21 +1,13 @@
 import { type ApolloClient } from '@apollo/client'
 import { type Redirect } from 'next'
 import {
-  ShopSessionCreatePartnerDocument,
-  type ShopSessionCreatePartnerMutation,
-  type ShopSessionCreatePartnerMutationVariables,
-} from '@/services/apollo/generated'
-import {
   type PageStory,
   type WidgetFlowStory,
   getStoryBySlug,
 } from '@/services/storyblok/storyblok'
 import { isWidgetFlowStory } from '@/services/storyblok/Storyblok.helpers'
 import { RoutingLocale } from '@/utils/l10n/types'
-import {
-  parseCustomerDataSearchParams,
-  parseShopSessionCreatePartnerSearchParams,
-} from './parseSearchParams'
+import { parseCustomerDataSearchParams } from './parseSearchParams'
 import { shopSessionCustomerUpdate } from './shopSessionCustomerUpdate'
 import { STORYBLOK_WIDGET_FOLDER_SLUG } from './widget.constants'
 
@@ -53,44 +45,6 @@ export const fetchFlowMetadata = async (
   }
 
   console.warn('Story is not a widget flow story:', story.slug)
-}
-
-type CreatePartnerShopSessionParams = {
-  apolloClient: ApolloClient<unknown>
-  searchParams: URLSearchParams
-} & Pick<ShopSessionCreatePartnerMutationVariables['input'], 'countryCode' | 'partnerName'>
-
-export const createPartnerShopSession = async (
-  params: CreatePartnerShopSessionParams,
-): Promise<[string, URLSearchParams]> => {
-  const [variables, unusedSearchParams] = parseShopSessionCreatePartnerSearchParams(
-    params.searchParams,
-  )
-
-  if (variables.externalMemberId) {
-    console.info(`Widget | Creating Trial Extension Shop Session: ${variables.externalMemberId}`)
-  } else {
-    console.info(`Widget | Creating Partner Shop Session`)
-  }
-
-  const result = await params.apolloClient.mutate<
-    ShopSessionCreatePartnerMutation,
-    ShopSessionCreatePartnerMutationVariables
-  >({
-    mutation: ShopSessionCreatePartnerDocument,
-    variables: {
-      input: {
-        countryCode: params.countryCode,
-        partnerName: params.partnerName,
-        initiatedFrom: 'WIDGET',
-        ...variables,
-      },
-    },
-  })
-
-  if (!result.data) throw new Error('Failed to create Partner Shop Session')
-
-  return [result.data.shopSessionCreatePartner.id, unusedSearchParams]
 }
 
 type UpdateCustomerDataIfPresentParams = {
