+++ apps/store/src/features/widget/createPartnerShopSession.ts
@@ -1,0 +1,69 @@
+import { type ApolloClient } from '@apollo/client'
+import { type VariablesOf } from '@graphql-typed-document-node/core'
+import { graphql } from '@/services/apollo/gql'
+import { SearchParam } from './parseSearchParams'
+
+const ShopSessionCreatePartnerDocument = graphql(/* GraphQL */ `
+  mutation ShopSessionCreatePartner($input: ShopSessionCreatePartnerInput!) {
+    shopSessionCreatePartner(input: $input) {
+      id
+    }
+  }
+`)
+
+type MutationVariables = VariablesOf<typeof ShopSessionCreatePartnerDocument>['input']
+
+type CreatePartnerShopSessionParams = {
+  apolloClient: ApolloClient<unknown>
+  searchParams: URLSearchParams
+} & Pick<MutationVariables, 'countryCode' | 'partnerName'>
+
+export const createPartnerShopSession = async (
+  params: CreatePartnerShopSessionParams,
+): Promise<[string, URLSearchParams]> => {
+  const [variables, unusedSearchParams] = parseShopSessionCreatePartnerSearchParams(
+    params.searchParams,
+  )
+
+  if (variables.externalMemberId) {
+    console.info(`Widget | Creating Trial Extension Shop Session: ${variables.externalMemberId}`)
+  } else {
+    console.info(`Widget | Creating Partner Shop Session`)
+  }
+
+  const result = await params.apolloClient.mutate({
+    mutation: ShopSessionCreatePartnerDocument,
+    variables: {
+      input: {
+        countryCode: params.countryCode,
+        partnerName: params.partnerName,
+        initiatedFrom: 'WIDGET',
+        ...variables,
+      },
+    },
+  })
+
+  if (!result.data) throw new Error('Failed to create Partner Shop Session')
+
+  return [result.data.shopSessionCreatePartner.id, unusedSearchParams]
+}
+
+const parseShopSessionCreatePartnerSearchParams = (
+  searchParams: URLSearchParams,
+): [Partial<MutationVariables>, URLSearchParams] => {
+  const updatedSearchParams = new URLSearchParams(searchParams.toString())
+
+  const externalRequestId = updatedSearchParams.get(SearchParam.ExternalRequestId)
+  if (externalRequestId) updatedSearchParams.delete(SearchParam.ExternalRequestId)
+
+  const externalMemberId = updatedSearchParams.get(SearchParam.ExternalMemberId)
+  if (externalMemberId) updatedSearchParams.delete(SearchParam.ExternalMemberId)
+
+  return [
+    {
+      ...(externalRequestId && { partnerRequestId: externalRequestId }),
+      ...(externalMemberId && { externalMemberId }),
+    },
+    updatedSearchParams,
+  ]
+}
